requirements.txt
django>=4.2.0
django-cors-headers>=4.0.0
python-dotenv>=1.0.0
openai>=1.0.0 